Add annotation tool for grouping detections into tracks.
The new tool has two inputs: image/video dataset and a detection
dataset. It produces a new detection dataset where TrackIDs are set.
The user clicks on sequences of detections over successive frames
that are associated with the same physical object.

This commit also adds a generic annotation UI interface for annotation
on image/video data. It includes components to deal with fetching a new
item (image or video) and, when labeling video, going between frames.
The annotate-shape tool is modified to use this generic interface, along
with the new detection-to-track tool.