Fix pytorch issue with non-materialized input dataset.
We do not have a way to load items from input dataset with non-default
item provider from the Python library. So previously pytorch training
would fail on such inputs.

Now we support a mechanism for ExecOps to specify a subgraph of other
operations that must be performed before running the ExecOp. For pytorch,
that means it will run new materialize operation on each input that
has non-default item provider to create a materialized dataset.

- Add VirtualNode and Runnable so that we can run ExecNodes that aren't
real ExecNodes in the database.
- Add ExecOp.Virtualize to produce a VirtualNode from an ExecNode.
- Add ExecOp.Resolve to optionally return a modified subgraph given a
VirtualNode, in case the op wants to run more operations that the
currently defined one. This provides the ExecOp with dynamic control
over the execution graph, which is going to complicate things but
hopefully will end up making sense.
- Add materialize operation, which takes an input dataset and copies
it to a new dataset that has default item provider (no referenced
items or virtual functions).
- Pytorch implements ExecOp.Resolve to materialize any inputs as needed.