Provide button for users to stop training jobs early.
Usually, when a job is terminated, its output datasets are left marked
incomplete so the node will be re-executed when a downstream node is
requested. With the new "Stop Training and Mark Done" button, the
outputs will be marked done. This way, training can be stopped early
if a user manually requests it, and the currently saved model can be
used immediately in downstream operations.

This is not really recommended since it makes the pipeline less
reproducible; but it is probably important for users to be able to
choose to not wait hours for a training job to complete if they are
already satisfied with the accuracy.

This commit also fixes tooltips.