Major updates to the Data API.
- The Data/DataImpl interfaces are completely reworked to make it more clear
what data is and what metadata is. Data itself is now an arbitrary interface{},
and reading and writing of data is handled by a new DataSpec interface.
- The old python operation is removed in favor of pythonv2 (which is now renamed "python").
- Resolves #29 since metadata is no longer embedded inside of the data objects.
- There is also now limited support for storing metadata with the dataset instead
of the item. There are some issues with this that'll need to be resolved later
though, and it's also not yet used by any operations.